neurons/validator.py: implement skew factor to convert win rate to weights
With the one-winner-per-sample model scoring, the win rate is a
normalized vector, i.e. its sum is 1 by definition. The win rate is
mapped to weights using a transformation.

The approach chosen is to raise each win rate to the power of
WEIGHT_SKEW_FACTOR, and normalized afterwards.

Lower win rates still receive a reward, but not proportional to their
win rate. This is to reward models that score low losses across a wide
range of samples.

Although this is a global constant, it is envisioned that when separate
competitions are added, the weight skew factor may be chosen differently
per competition.