hugging_face_model_store.py: check locks before starting download
huggingface_hub uses filelock.FileLock to lock downloaded files. This
may lead to stalled processes if, for whatever reason, another process
holds a lock. huggingface_hub doesn't implement a timeout so the only
way to handle this is either running it in a subprocess (suboptimal) or
checking whether any locks are active, before starting download.