neurons/validation.py: fix compute_wins()
Sample scores were only compared to newer models, which led to the
possibility that an older model A would not win from a newer, better
model C, that would subsequently lose against a model B posted between
model A and C, that need not be better than model A.

This is fixed by testing whether models are better in absolute sense
than older models. So in the example above, if model A doesn't win
due to model C, model B doesn't win because it is not better than
model A, and model C will win.